Implements a feature to show live games and their current viewer number.

+++ wsgateway/ws/manage.go
@@ -89,14 +89,19 @@ func (s *Server) manageSessionsState() {
 			// remove list of viewers of games that ended
 			// it's better to be done in only one instance of wsGateway with a master/slave mechanism!
 			gamesId := s.em.getAll()
-			err := s.cache.removeGamesViewersLists(context.Background(), gamesId)
-			if err != nil {
-				s.l.Error(err.Error())
-			} else {
-				s.em.remove(gamesId)
+			if len(gamesId) > 0 {
+				err := s.cache.removeGamesViewersLists(context.Background(), gamesId)
+				if err != nil {
+					s.l.Error(err.Error())
+				} else {
+					s.em.remove(gamesId)
+				}
 			}
 
 			ss := s.sm.getAll()
+			if len(ss) == 0 {
+				continue
+			}
 			// update sessions timestamp in redis cache, every 1 minute
 			// this should be done in every wsGateway instances
 			if err := s.cache.updateSessionsTimestamp(context.Background(), ss...); err != nil {
